Refactor tear sheet to tables
